refactor: constants in index
