+++ src/components/molecules/CheckoutForm/useCheckoutForm.ts
@@ -1,8 +1,7 @@
 import { useStripe, useElements } from '@stripe/react-stripe-js'
 import { type FormEvent, useState, type Dispatch, type SetStateAction } from 'react'
 import './CheckoutForm.scss'
-import APP_ERRORS from '@/constants/appErrors'
-import { ROUTES } from '@/constants'
+import { ROUTES, APP_ERRORS } from '@/constants'
 import type { StripeAddress } from '@/types/order'
 import { useCustomerContext } from '@/providers/customer/CustomerContext'
 import { type Stripe, type StripeElements } from '@stripe/stripe-js'
