+++ src/components/molecules/PaymentStatus/PaymentStatus.tsx
@@ -1,8 +1,7 @@
 import { type FC, type ReactElement } from 'react'
 import Alert from 'react-bootstrap/Alert'
 import Button from 'react-bootstrap/Button'
-import FORMS from '@/constants/forms'
-import { ROUTES } from '@/constants'
+import { ROUTES, FORMS } from '@/constants'
 import Loading from '@/components/atoms/Loading/Loading'
 import usePaymentStatus from './usePaymentStatus'
 import { useNavigate } from 'react-router-dom'
